backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
sqlalchemy>=2.0.35
alembic>=1.13.3
pydantic>=2.9.2
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
httpx>=0.27.2
ddgs>=9.0.0
trafilatura>=1.12.0
beautifulsoup4>=4.12.3
bleach>=6.2.0
sentence-transformers>=3.2.0
scikit-learn>=1.5.0
apscheduler>=3.10.4
Pillow>=10.4.0
python-slugify>=8.0.4
aiosqlite>=0.20.0
pydantic[email]
